src/requirements.txt
fastapi
PyJWT
uvicorn
cryptography
python-dotenv
